LayoutTests/fast/events/pointer/ios/drag-populates-pointer-events-with-movementxy-fields-expected.txt

Test that the movementX/Y fields of the pointer events fired from a touch drag are correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



    pointerdown fired.
PASS: movementX is correct.
PASS: movementY is correct.
pointermove fired.
PASS: movementX is correct.
PASS: movementY is correct.
pointermove fired.
PASS: movementX is correct.
PASS: movementY is correct.
pointermove fired.
PASS: movementX is correct.
PASS: movementY is correct.
pointermove fired.
PASS: movementX is correct.
PASS: movementY is correct.
pointermove fired.
PASS: movementX is correct.
PASS: movementY is correct.
pointerup fired.
PASS: movementX is correct.
PASS: movementY is correct.
PASS successfullyParsed is true

TEST COMPLETE
